Refactor: Standardize typing, docstrings, and English translation
- Enforce explicit type hinting in class attributes and methods.
- Translate all code comments and docstrings to English.
- Adopt Google-style docstrings across the project.
- Fix syntax for equality comparisons and membership checks.
- Improve code formatting for method arguments.

+++ src/operations_manager.py
@@ -27,7 +27,7 @@ class Operation:
             type: str,
             cash_amount: float,
             ticker: str,
-            stock_price: int,
+            stock_price: float,
             successful: bool,
             date: str,
             trigger: str = "manual"
